requirements-test.txt
aioresponses==0.7.4
black==23.1.0
pytest-cov==4.0.0
flake8==6.0.0
flake8-docstrings==1.7.0
flake8-noqa==1.3.0
isort==5.12.0
mypy==1.1.1
pydocstyle==6.3.0
pytest==7.2.2
pytest-aiohttp==1.0.4
ruff==0.0.254
types-orjson==3.6.2
